Make popup component reusable (#67)
